# backend/docs/swagger.yaml
basePath: /api/v1
definitions:
  domain.AuthLoginRequest:
    properties:
      email:
        type: string
      password:
        type: string
    required:
    - email
    - password
    type: object
  domain.AuthRegisterRequest:
    properties:
      email:
        type: string
      name:
        maxLength: 20
        minLength: 4
        type: string
      password:
        maxLength: 100
        minLength: 8
        type: string
    required:
    - email
    - name
    - password
    type: object
  domain.Claims:
    properties:
      user_id:
        type: integer
    type: object
  domain.Metadata:
    properties:
      limit:
        type: integer
      order:
        enum:
        - asc
        - desc
        type: string
      page:
        type: integer
      sort:
        enum:
        - id
        - user_id
        - name
        - title
        - created_at
        - updated_at
        type: string
      total_pages:
        type: integer
      total_records:
        type: integer
    type: object
  domain.NoteAuthor:
    properties:
      avatar_url:
        type: string
      id:
        type: integer
      name:
        type: string
    type: object
  domain.NotePaginationResponse:
    properties:
      metadata:
        $ref: '#/definitions/domain.Metadata'
      notes:
        items:
          $ref: '#/definitions/domain.NoteResponse'
        type: array
    type: object
  domain.NoteRequest:
    properties:
      content:
        type: string
      cover_url:
        type: string
      description:
        maxLength: 50
        type: string
      id:
        type: integer
      title:
        maxLength: 25
        type: string
      user_id:
        type: integer
      visibility:
        enum:
        - private
        - public
        type: string
    required:
    - content
    - cover_url
    - description
    - title
    - visibility
    type: object
  domain.NoteResponse:
    properties:
      author:
        $ref: '#/definitions/domain.NoteAuthor'
      content:
        type: string
      cover_url:
        type: string
      created_at:
        type: string
      description:
        type: string
      id:
        type: integer
      title:
        type: string
      updated_at:
        type: string
      visibility:
        type: string
    type: object
  domain.NoteUpdateRequest:
    properties:
      content:
        type: string
      cover_url:
        type: string
      description:
        maxLength: 50
        type: string
      id:
        type: integer
      title:
        maxLength: 25
        type: string
      user_id:
        type: integer
      visibility:
        enum:
        - private
        - public
        type: string
    type: object
  domain.UserPaginationResponse:
    properties:
      metadata:
        $ref: '#/definitions/domain.Metadata'
      users:
        items:
          $ref: '#/definitions/domain.UserResponse'
        type: array
    type: object
  domain.UserRequest:
    properties:
      avatar_url:
        type: string
      bio:
        maxLength: 50
        type: string
      email:
        type: string
      id:
        type: integer
      name:
        maxLength: 20
        minLength: 4
        type: string
      password:
        maxLength: 100
        minLength: 8
        type: string
    type: object
  domain.UserResponse:
    properties:
      avatar_url:
        type: string
      bio:
        type: string
      created_at:
        type: string
      id:
        type: integer
      name:
        type: string
      tokens:
        $ref: '#/definitions/domain.UserToken'
      updated_at:
        type: string
    type: object
  domain.UserToken:
    properties:
      access_token:
        type: string
      claims:
        $ref: '#/definitions/domain.Claims'
      refresh_token:
        type: string
    type: object
info:
  contact: {}
  description: golang crud api
  title: Zenith
  version: "1.0"
paths:
  /auth/login:
    post:
      consumes:
      - application/json
      description: Log in an existing user with the provided email and password
      parameters:
      - description: User login request object
        in: body
        name: user
        required: true
        schema:
          $ref: '#/definitions/domain.AuthLoginRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Successfully logged in
          schema:
            $ref: '#/definitions/domain.UserResponse'
      summary: User login
      tags:
      - auth
  /auth/logout:
    post:
      description: Log out the currently logged-in user
      produces:
      - application/json
      responses:
        "200":
          description: Successfully logged out
      summary: User logout
      tags:
      - auth
  /auth/refresh:
    post:
      description: Refresh the access token using the refresh token
      produces:
      - application/json
      responses:
        "200":
          description: Successfully refreshed token
          schema:
            $ref: '#/definitions/domain.UserToken'
      summary: Refresh access token
      tags:
      - auth
  /auth/register:
    post:
      consumes:
      - application/json
      description: Register a new user with the specified name, email, and password
      parameters:
      - description: User registration request object
        in: body
        name: user
        required: true
        schema:
          $ref: '#/definitions/domain.AuthRegisterRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Successfully registered a new user
          schema:
            $ref: '#/definitions/domain.UserResponse'
      summary: Register a new user
      tags:
      - auth
  /notes:
    get:
      description: Retrieve all available notes
      parameters:
      - description: Filter notes by title
        in: query
        name: title
        type: string
      - description: Filter notes by author
        in: query
        name: author
        type: string
      - description: Filter notes by user ID
        in: query
        name: user_id
        type: string
      - description: Filter notes by visibility
        in: query
        name: visibility
        type: string
      - description: Sorting (e.g., +title, -created_at)
        in: query
        name: sort
        type: string
      - description: Sort order (e.g., asc, desc)
        in: query
        name: order
        type: string
      - description: Page number
        in: query
        name: page
        type: integer
      - description: Number of items per page
        in: query
        name: limit
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Successfully retrieved all notes
          schema:
            $ref: '#/definitions/domain.NotePaginationResponse'
      summary: Get all notes
      tags:
      - note
    post:
      consumes:
      - application/json
      description: Create a new note with the specified title, content, and visibility
      parameters:
      - description: Note request object
        in: body
        name: note
        required: true
        schema:
          $ref: '#/definitions/domain.NoteRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Successfully created a new note
          schema:
            $ref: '#/definitions/domain.NoteResponse'
      summary: Create a new note
      tags:
      - note
  /notes/{id}:
    delete:
      description: Delete an existing note based on the provided ID
      parameters:
      - description: Note ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Successfully deleted a note by ID
      summary: Delete a note by ID
      tags:
      - note
    get:
      consumes:
      - application/json
      description: Retrieve a note based on the provided ID
      parameters:
      - description: Note ID
        in: path
        name: id
        required: true
        type: integer
      - description: Page number
        in: query
        name: page
        type: integer
      - description: Number of items per page
        in: query
        name: limit
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Successfully retrieved a note by ID
          schema:
            $ref: '#/definitions/domain.NoteResponse'
      summary: Get a note by ID
      tags:
      - note
    put:
      consumes:
      - application/json
      description: Update an existing note based on the provided ID
      parameters:
      - description: Note ID
        in: path
        name: id
        required: true
        type: integer
      - description: Updated note object
        in: body
        name: note
        required: true
        schema:
          $ref: '#/definitions/domain.NoteUpdateRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Successfully updated a note by ID
          schema:
            $ref: '#/definitions/domain.NoteResponse'
      summary: Update a note by ID
      tags:
      - note
  /users:
    get:
      description: Retrieve data of all registered users
      parameters:
      - description: Filter users by ID
        in: query
        name: id
        type: integer
      - description: Filter users by name
        in: query
        name: name
        type: string
      - description: Get users details
        in: query
        name: details
        type: boolean
      - description: Sorting (e.g., +name, -created_at)
        in: query
        name: sort
        type: string
      - description: Sort order (e.g., asc, desc)
        in: query
        name: order
        type: string
      - description: Page number
        in: query
        name: page
        type: integer
      - description: Number of items per page
        in: query
        name: limit
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Successfully retrieved all user data
          schema:
            items:
              $ref: '#/definitions/domain.UserPaginationResponse'
            type: array
      summary: Get all users
      tags:
      - user
  /users/{id}:
    delete:
      description: Delete an existing user based on the provided ID
      parameters:
      - description: User ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Successfully deleted user by ID
      summary: Delete a user by ID
      tags:
      - user
    get:
      description: Retrieve data of a user based on the provided ID
      parameters:
      - description: User ID
        in: path
        name: id
        required: true
        type: integer
      - description: Page number
        in: query
        name: page
        type: integer
      - description: Number of items per page
        in: query
        name: limit
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Successfully retrieved user by ID
          schema:
            $ref: '#/definitions/domain.UserResponse'
      summary: Get a user by ID
      tags:
      - user
    put:
      consumes:
      - application/json
      description: Update data of an existing user based on the provided ID
      parameters:
      - description: User ID
        in: path
        name: id
        required: true
        type: integer
      - description: Updated user data object
        in: body
        name: user
        required: true
        schema:
          $ref: '#/definitions/domain.UserRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Successfully updated user by ID
          schema:
            $ref: '#/definitions/domain.UserResponse'
      summary: Update user data by ID
      tags:
      - user
  /users/me:
    get:
      description: Retrieve information of the currently authenticated user
      produces:
      - application/json
      responses:
        "200":
          description: Successfully retrieved current user's information
          schema:
            $ref: '#/definitions/domain.UserResponse'
      summary: Get current user's information
      tags:
      - user
swagger: "2.0"